* Ommited unused imports in following files: 	* AmazonS3ErrorUtils.java 	* AmazonS3PluginTest.java 	* AmazonS3ErrorUtilsTest.java
* Moved test functions
* combined instance checking for AmazonServiceException and its subclass
